backend/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.6
email-validator==2.2.0
